Add default value
still some unit test to fix
